imapserver: redo the connection pool
We were using more lists then we needed (e.g. self.lastowned and
self.availableconnections) and the code was not extremely well
documented. This commit simplies some code and adds more comments on
what we are really doing.

I seriously doubt that the 'optimization' of trying to pick an imap
connection that belonged to the same thread before that is was (and still is)
here is very effective. First, there should usually not be that many
unused availableconnections be around that we can pick one out of many,
most of the time there will be only one available anyway. Second, do we
initiate new threads for e.g. the copying of each message, so the same
thread id will invoking acquireconnection() twice. Anyway, it doesn't
really hurt either, so I am letting it in for now.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>