Rework the threadutil stuff
Do away with a whole lot of complexity in monitoring exited threads that
we were not using anyway. The new OfflineImapError levels should be able
to give us the same thing.

The only downside is that rather than listening for exited threads we
sleep between checks for 0.5 seconds, so the multi-thread case
takes a bit longer on average to exit. We can always tweak this value if
we find that unacceptable.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>