Automagically detect the trash and spam folders
Rather than hardcoding the trash and spam folders by hand, which is
complicated and error pronem, we detect it from the Gmail server. It
takes less than 1 second and needs to be done once per sync run. We only
do this if the user has configured "realdelete=True".

If it is deemed to costly in terms of performance, we should cache
the automatically detected value somewhere rather than having the user
to find out if the correct value is [Gmail]/Trash, [Google Mail]/Trash,
or [Gmail]/Papierkorb or whatever.

We need this information as we need to move a mail into this folder in
order to really delete it from the Gmail account (removing it from a
folder will simply remove a tag in Gmail and keep the mail in All Mails.

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>