Rework UID Mapped Folder implementation
The Folder implementation needed some love. It used more dicts and lists
than was necessary (self.disk_l2r, self.disk_r2l, self.r2l) and which
could be done away with. Also there is no need to invoke a separate
IMAPFolder() and store it in self._mb we can use out instance just
fine. All functions in IMAPFolder() which are passed in some uid need to
be overriden though, to make sure that we translate RUID (remote uid) to
the LUID (local uid).

Signed-off-by: Sebastian Spaeth <Sebastian@SSpaeth.de>